Optemized profile picture loading
Before, every tune a profile picture needed to be loaded a new image resource was created even if it is the same image for the same user.  (I don't actually know if this is true. Maybe Godot detects when a duplicate image resource is made) Now, the peer service keeps a list of image resources linked to each peer's id which can be pulled whenever an element needs to display a peer's profile picture.